+++ src/client/js/components/Admin/App/SesSetting.jsx
@@ -2,34 +2,16 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { withTranslation } from 'react-i18next';
-import loggerFactory from '@alias/logger';
 
 import { withUnstatedContainers } from '../../UnstatedUtils';
-import { toastSuccess, toastError } from '../../../util/apiNotification';
 import { withLoadingSppiner } from '../../SuspenseUtils';
 
 
 import AppContainer from '../../../services/AppContainer';
 import AdminAppContainer from '../../../services/AdminAppContainer';
 
-const logger = loggerFactory('growi:smtpSettings');
-
-
 function SmtpSetting(props) {
-  const { adminAppContainer, t } = props;
-
-  async function submitHandler() {
-    const { t } = props;
-
-    try {
-      await adminAppContainer.updateSesSettingHandler();
-      toastSuccess(t('toaster.update_successed', { target: t('admin:app_setting.ses_settings') }));
-    }
-    catch (err) {
-      toastError(err);
-      logger.error(err);
-    }
-  }
+  const { adminAppContainer } = props;
 
   return (
     <React.Fragment>
@@ -64,14 +46,6 @@ function SmtpSetting(props) {
                 adminAppContainer.changeSesSecretAccessKey(e.target.value);
               }}
             />
-          </div>
-        </div>
-
-        <div className="row my-3">
-          <div className="mx-auto">
-            <button type="button" className="btn btn-primary" onClick={submitHandler} disabled={adminAppContainer.state.retrieveError != null}>
-              { t('Update') }
-            </button>
           </div>
         </div>
       </div>
